backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
alembic==1.16.5
pymysql==1.1.1
redis==6.4.0
PyJWT==2.10.1
passlib[bcrypt]==1.7.4
pydantic-settings==2.10.1
python-dotenv==1.1.1
